Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/sandbox/compress/trunk@749856 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/compress/archivers/tar/TarArchiveInputStream.java
@@ -286,7 +286,7 @@ public class TarArchiveInputStream extends ArchiveInputStream {
      */
     public int read() throws IOException {
         int num = read(oneBuf, 0, 1);
-        return num == -1 ? -1 : ((int) oneBuf[0]) & BYTE_MASK;
+        return num == -1 ? -1 : oneBuf[0] & BYTE_MASK;
     }
 
     /**
